+++ CotEditor/Sources/SidebarViewController.swift
@@ -63,13 +63,11 @@ final class SidebarViewController: NSTabViewController {
         
         super.viewDidLoad()
         
-        // bind segmentedControl manually  (2016-09 on macOS 10.12)
-        if let segmentedControl = (self.tabView as? InspectorTabView)?.segmentedControl {
-            segmentedControl.bind(.selectedIndex, to: self, withKeyPath: #keyPath(selectedTabViewItemIndex))
-        }
-        
         // select last used pane
         self.selectedTabViewItemIndex = UserDefaults.standard[.selectedInspectorPaneIndex]
+        
+        // bind segmentedControl manually  (2016-09 on macOS 10.12)
+        (self.tabView as! InspectorTabView).segmentedControl.bind(.selectedIndex, to: self, withKeyPath: #keyPath(selectedTabViewItemIndex))
         
         // set accessibility
         self.view.setAccessibilityElement(true)