+++ CotEditor/Sources/StatusBarController.swift
@@ -85,11 +85,14 @@ final class StatusBarController: NSViewController {
     
     
     /// request analyzer to update editor info
-    override func viewDidAppear() {
+    override func viewWillAppear() {
         
-        super.viewDidAppear()
+        super.viewWillAppear()
+        
+        assert(self.documentAnalyzer != nil)
         
         self.documentAnalyzer?.needsUpdateStatusEditorInfo = true
+        self.documentAnalyzer?.invalidateEditorInfo()
     }
     
     
@@ -120,7 +123,9 @@ final class StatusBarController: NSViewController {
         didSet {
             guard let analyzer = documentAnalyzer else { return }
             
-            analyzer.needsUpdateStatusEditorInfo = !self.view.isHiddenOrHasHiddenAncestor
+            if self.isViewLoaded {
+                analyzer.needsUpdateStatusEditorInfo = !self.view.isHiddenOrHasHiddenAncestor
+            }
             
             NotificationCenter.default.addObserver(self, selector: #selector(updateEditorStatus), name: DocumentAnalyzer.didUpdateEditorInfoNotification, object: analyzer)
             NotificationCenter.default.addObserver(self, selector: #selector(updateDocumentStatus), name: DocumentAnalyzer.didUpdateFileInfoNotification, object: analyzer)
@@ -140,8 +145,11 @@ final class StatusBarController: NSViewController {
         
         assert(Thread.isMainThread)
         
-        guard !self.view.isHiddenOrHasHiddenAncestor else { return }
-        guard let info = self.documentAnalyzer?.info else { return }
+        guard
+            self.isViewLoaded,
+            !self.view.isHiddenOrHasHiddenAncestor,
+            let info = self.documentAnalyzer?.info
+            else { return }
         
         let appearance = self.view.effectiveAppearance
         let defaults = UserDefaults.standard
@@ -180,8 +188,11 @@ final class StatusBarController: NSViewController {
         
         assert(Thread.isMainThread)
         
-        guard !self.view.isHiddenOrHasHiddenAncestor else { return }
-        guard let info = self.documentAnalyzer?.info else { return }
+        guard
+            self.isViewLoaded,
+            !self.view.isHiddenOrHasHiddenAncestor,
+            let info = self.documentAnalyzer?.info
+            else { return }
         
         let defaults = UserDefaults.standard
         let status = NSMutableAttributedString()