+++ CHANGELOG.md
@@ -13,6 +13,7 @@ Change Log
 ### Fixes
 
 - Fix an issue where the Unicode character name for `U+FEFF` (ZERO WIDTH NO-BREAK SPACE) was empty.
+- Fix an issue where the application did crash when selected text contains some specific characters.
 - Fix an issue where needless live document analysis performed even when the status bar and inspector are invisible.
 - Fix missing localization.